+++ KarhooUISDK/CommonUI/Views/RoundedCorner.swift
@@ -1,0 +1,23 @@
+//
+//  RoundedCorner.swift
+//  KarhooUISDK
+//
+//  Created by Bartlomiej Sopala on 23/01/2023.
+//  Copyright © 2023 Flit Technologies Ltd. All rights reserved.
+//
+
+import Foundation
+import SwiftUI
+
+struct RoundedCorner: Shape {
+    var radius: CGFloat = .infinity
+    var corners: UIRectCorner = .allCorners
+    func path(in rect: CGRect) -> Path {
+        let path = UIBezierPath(
+            roundedRect: rect,
+            byRoundingCorners: corners,
+            cornerRadii: CGSize(width: radius, height: radius)
+        )
+        return Path(path.cgPath)
+    }
+}
